+++ Js2IL.Tests/Function/JavaScript/Function_Prototype_Constructor_ReferencesFunction.js
@@ -1,0 +1,3 @@
+"use strict";
+
+console.log(Function.prototype.constructor === Function);